Adds a sleep(5) and fixes access
The Access on a bot was previously deleted entirely when it completed a
call from the AI. The bot now retains its original access.

Also adds a small delay so that bots animate properly upon moving,
instead of appearing to teleport.

The "Call" button will lose its call link instead of disappearing
completely.