Further improvements!
AIs can now detect bots again while in a container! (Carded, on a MULE,
in an APC)

Made a tweak so paths are not duplicated in the bot's variables.

MULEs are working again! (MULE hacking to come in another PR)

Brought the ED-209 up to date.

Made the bot menu text stop generated the weird y character in names.