Access fixes and set AI call range to 150
I somehow broke access for trying to unlock certain bots. This sets it
to req_one_access so anyone with the appropriate access on their ID will
have access to the bot.

The AI's max A-star range is now set to 150 so bots can reach most areas
in the station without it being insane.