Patrol code overhaul + tweaks
Lots of changes in this one! First off, medibots and floorbots can now
be commanded to patrol. Access to the bot's settings is required to make
it patrol.

Patrol and summon code is centralized such that patrolling and summoning
can be given to all simple bots (anything that is not a MULE).

The AI can read the status of MULEs through its interface.

ED-209 can now be summoned by PDA as Securitrons can!