Ports a /TG/ camera menu improvement!
Ports: https://github.com/tgstation/-tg-station/pull/4203

This bugfix features a Cancel button to the tracking and camera list
input windows, allowing an AI to cancel a camera command.

It also re-adds the ability to track from the command line. Type "Track
Miku Frost" to stalk her relentlessly. (It will also auto-complete, so
you can type "Track Saiga P" and a space so you do not have to know how
to spell an entire name to input it).